external_modules/Tier_10_Clean/requirements.txt
ollama>=0.1.0
astor>=0.8.1
pytest>=7.0.0
pyyaml>=6.0
transformers>=4.37.0
torch>=2.0.0
accelerate>=0.26.0
peft>=0.7.0
einops
